+++ substrabac/substrapp/signals/model/post_delete.py
@@ -7,4 +7,5 @@ def model_post_delete(sender, instance, **kwargs):
     instance.file.delete(False)
 
     directory = path.join(getattr(settings, 'MEDIA_ROOT'), 'models/{0}'.format(instance.pk))
-    shutil.rmtree(directory)
+    if path.exists(directory):
+        shutil.rmtree(directory)
